+++ build.gradle
@@ -64,8 +64,6 @@ dependencies {
 	implementation 'org.apache.poi:poi:4.1.2'
 	implementation 'org.apache.poi:poi-ooxml:4.1.2'
 	implementation 'commons-io:commons-io:2.11.0'
-
-
 }
 
 tasks.named('test') {
