+++ src/main/java/umc/product/domain/member/repository/jpa/MemberJpaRepository.java
@@ -1,4 +1,4 @@
-package umc.product.domain.member.repository;
+package umc.product.domain.member.repository.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import umc.product.domain.member.entity.Member;
