+++ src/main/java/umc/product/domain/member/repository/jpa/MemberLoginInfoJpaRepository.java
@@ -1,10 +1,10 @@
-package umc.product.domain.member.repository;
+package umc.product.domain.member.repository.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import umc.product.domain.member.entity.MemberLoginInfo;
 
 import java.util.Optional;
 
-public interface MemberLoginInfoRepository extends JpaRepository<MemberLoginInfo, Long> {
+public interface MemberLoginInfoJpaRepository extends JpaRepository<MemberLoginInfo, Long> {
     Optional<MemberLoginInfo> findByMemberLoginId(String memberLoginId);
 }
