+++ src/main/java/umc/product/domain/member/repository/jpa/MemberOutJpaRepository.java
@@ -1,4 +1,4 @@
-package umc.product.domain.member.repository;
+package umc.product.domain.member.repository.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import umc.product.domain.member.entity.MemberOut;
@@ -6,7 +6,7 @@ import umc.product.domain.member.entity.MemberOut;
 import java.util.Optional;
 
 
-public interface MemberOutRepository extends JpaRepository<MemberOut, Long> {
+public interface MemberOutJpaRepository extends JpaRepository<MemberOut, Long> {
     Optional<MemberOut> findMemberOutById(Long outId);
 }
