+++ src/main/java/umc/product/domain/member/strategy/impl/AnonymousLoginStrategy.java
@@ -8,15 +8,18 @@ import umc.product.domain.member.entity.Member;
 import umc.product.domain.member.entity.enums.Role;
 import umc.product.domain.member.dto.response.member.auth.MemberLoginResponse;
 import umc.product.domain.member.mapper.MemberMapper;
-import umc.product.domain.member.repository.MemberRepository;
+import umc.product.domain.member.repository.querydsl.MemberRepository;
 import umc.product.domain.member.serviceImpl.member.MemberServiceImpl;
 import umc.product.domain.member.strategy.LoginStrategy;
+import umc.product.global.common.exception.RestApiException;
 import umc.product.global.config.security.jwt.JwtProvider;
 import umc.product.global.config.security.jwt.TokenInfo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 import java.util.Optional;
+
+import static umc.product.domain.member.status.MemberErrorStatus.NOT_SUPPORT_LOGIN_TYPE;
 
 @RequiredArgsConstructor
 @Component
@@ -46,8 +49,7 @@ public class AnonymousLoginStrategy implements LoginStrategy {
 
     @Override
     public MemberLoginResponse login(AdminLoginRequest request) {
-        // todo : MemberLoginRequest  방식은 지원하지 않습니다. RestApiException으로 변경
-        throw new UnsupportedOperationException("MemberLoginRequest  방식은 지원하지 않습니다.");
+        throw new RestApiException(NOT_SUPPORT_LOGIN_TYPE);
     }
 
     private MemberLoginResponse saveNewMember(String clientId, LoginType loginType) {
